+++ src/app/shared/components/app-footer/app-footer.component.ts
@@ -1,13 +1,13 @@
 import { CommonModule } from '@angular/common';
-import { Component, inject } from '@angular/core';
+import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 
 import { AppService } from '../../../core/services/app.service';
 
 @Component({
   selector: 'app-footer',
-  standalone: true,
   imports: [CommonModule],
   templateUrl: 'app-footer.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class AppFooterComponent {
   private appService = inject(AppService);